made lifeline check size of message name when marking a file as deleted removed lifeline-msgs rsync into its own script which is now run hourly from the main server

+++ trunk/asterisk-common/agi-bin/Lifeline.pm
@@ -383,7 +383,7 @@ sub del_all {
 sub clean_up_msgs {
 	my $ll = shift;
 	foreach (@{$ll->{msgs}->{list}}) {
-		if ($_->{deleted}) {
+		if ($_->{deleted} and $_->{msg} ne '') {
 			my $mname = $_->{msg}.".".$ll->{rectype};
 			my $dname = $_->{msg}.".deleted.".$ll->{rectype};
 			unlink $dname if -f $dname;
